v6

+++ src/main/java/com/repetition/rest_api/service/UserService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -17,7 +18,8 @@ public class UserService {
     }
     // wyszykaj użytkownika po id
     public User getUser(long id){
-        return userRepository.getOne(id);
+        Optional<User> isUser = userRepository.findById(id);
+        return isUser.orElseGet(() -> new User(0, "B/D"));
     }
     // utwórz nowego użytkownika
     public void createUser(User user){
@@ -29,6 +31,9 @@ public class UserService {
     }
     // usuń użytkownika po id
     public void deleteUser(long id){
-        userRepository.deleteById(id);
+        Optional<User> isUser = userRepository.findById(id);
+        if(isUser.isPresent()) {
+            userRepository.deleteById(id);
+        }
     }
 }
